Avoid warning about type-truncation.

+++ src/libxs_malloc.c
@@ -1326,7 +1326,7 @@ LIBXS_API void libxs_free(const void* memory)
 
       /* check if memory belongs to scratch domain or local domain */
       if (NULL != info && pool->instance.buffer <= buffer && buffer < (pool->instance.buffer + info->size)) {
-        const size_t counter = LIBXS_ATOMIC_SUB_FETCH(&pool->instance.counter, (size_t)1, LIBXS_ATOMIC_SEQ_CST);
+        const size_t counter = LIBXS_ATOMIC_SUB_FETCH(&pool->instance.counter, 1, LIBXS_ATOMIC_SEQ_CST);
 
         assert(pool->instance.buffer <= pool->instance.head);
         if (0 == counter) { /* reallocate scratch domain */
